rename rds to postgres, update env vars
