+++ apps/web/src/utils/proofs/proofs_storage.ts
@@ -1,4 +1,4 @@
-import { getDb } from 'apps/web/src/utils/datastores/rds';
+import { getDb } from 'apps/web/src/utils/datastores/postgres';
 import { Address } from 'viem';
 
 export enum ProofTableNamespace {
